+++ src/zrb/util/git.py
@@ -1,16 +1,11 @@
 import os
 from collections.abc import Callable
-from typing import Any
-
-from pydantic import BaseModel
+from typing import Any, TYPE_CHECKING
 
 from zrb.util.cmd.command import run_command
 
-
-class DiffResult(BaseModel):
-    created: list[str]
-    removed: list[str]
-    updated: list[str]
+if TYPE_CHECKING:
+    from zrb.util.git_diff_result import DiffResult
 
 
 async def get_diff(
@@ -18,7 +13,7 @@ async def get_diff(
     source_commit: str,
     current_commit: str,
     print_method: Callable[..., Any] = print,
-) -> DiffResult:
+) -> "DiffResult":
     """
     Get the difference between two commits in a Git repository.
 
@@ -34,6 +29,8 @@ async def get_diff(
     Raises:
         Exception: If the git command returns a non-zero exit code.
     """
+    from zrb.util.git_diff_result import DiffResult
+
     cmd_result, exit_code = await run_command(
         cmd=["git", "diff", source_commit, current_commit],
         cwd=repo_dir,
