+++ src/zrb/util/git_diff_result.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class DiffResult(BaseModel):
+    created: list[str]
+    removed: list[str]
+    updated: list[str]
