+++ src/zrb/util/git_subtree_config.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+
+
+class SingleSubTreeConfig(BaseModel):
+    repo_url: str
+    branch: str
+    prefix: str
+
+
+class SubTreeConfig(BaseModel):
+    data: dict[str, SingleSubTreeConfig]
